cosmetic: put tool bar above everything
