+++ src/pages/settings/users/UserList.jsx
@@ -1,20 +1,9 @@
 import { useState, useMemo, useRef } from 'react'
-import { toast } from 'react-toastify'
 import { DataTable } from 'primereact/datatable'
 import { Column } from 'primereact/column'
 import { ContextMenu } from 'primereact/contextmenu'
-import { confirmDialog, ConfirmDialog } from 'primereact/confirmdialog'
-import { TablePanel, Button, Section, Toolbar, InputText } from '@ynput/ayon-react-components'
-// Comps
-import NewUserDialog from './newUserDialog'
-import SetPasswordDialog from './SetPasswordDialog'
-import RenameUserDialog from './RenameUserDialog'
-// utils
-import axios from 'axios'
+import { TablePanel, Section } from '@ynput/ayon-react-components'
 import './users.sass'
-import useSearchFilter from '/src/hooks/useSearchFilter'
-import { useGetUsersQuery } from '../../../services/user/getUsers'
-import { useGetRolesQuery } from '/src/services/getRoles'
 
 const buildUserDetailData = (projectNames, roleNames, users, lastSelectedUser) => {
   let roles = []
@@ -50,50 +39,26 @@ const buildUserDetailData = (projectNames, roleNames, users, lastSelectedUser) =
   }
 }
 
-// TODO: Remove classname assignments and do in styled components
-const formatRoles = (rowData, selectedProjects) => {
-  let res = {}
-  if (rowData.isAdmin) res.admin = { cls: 'role admin' }
-  else if (rowData.isService) res.service = { cls: 'role manager' }
-  else if (rowData.isManager) res.manager = { cls: 'role manager' }
-  else if (!selectedProjects) {
-    for (const name of rowData.defaultRoles || []) res[name] = { cls: 'role default' }
-  } else {
-    const roleSet = JSON.parse(rowData.roles)
-    for (const projectName of selectedProjects) {
-      for (const roleName of roleSet[projectName] || []) {
-        if (roleName in res) res[roleName].count += 1
-        else res[roleName] = { count: 1 }
-        res[roleName].cls =
-          res[roleName].count === selectedProjects.length ? 'role all' : 'role partial'
-      }
-    }
-  }
-
-  return { ...rowData, roles: res, rolesList: Object.keys(res) }
-}
-
-const UserList = ({ selectedProjects, selectedUsers, onSelectUsers, setUserDetailData }) => {
-  const [showNewUser, setShowNewUser] = useState(false)
+const UserList = ({
+  selectedProjects,
+  selectedUsers,
+  setSelectedUsers,
+  setUserDetailData,
+  rolesList,
+  userList,
+  tableList,
+  setShowRenameUser,
+  setShowSetPassword,
+  onDelete,
+  isLoading,
+  isLoadingRoles,
+}) => {
   const [lastSelectedUser, setLastSelectedUser] = useState(null)
-  const [showRenameUser, setShowRenameUser] = useState(false)
-  const [showSetPassword, setShowSetPassword] = useState(false)
   const contextMenuRef = useRef(null)
-
-  const { data: userList = [], isLoading, isError, isFetching, refetch } = useGetUsersQuery()
-  if (isError) toast.error('Unable to load users')
-
-  const {
-    data: rolesList = [],
-    isLoading: isLoadingRoles,
-    isError: isErrorRoles,
-  } = useGetRolesQuery()
-  if (isErrorRoles) toast.error('Unable to load roles')
 
   // Selection
 
   const selection = useMemo(() => {
-    if (isFetching) return []
     let result = []
     let lastUsr = null
     for (const user of userList) {
@@ -105,35 +70,13 @@ const UserList = ({ selectedProjects, selectedUsers, onSelectUsers, setUserDetai
       setUserDetailData(buildUserDetailData(selectedProjects, rolesList, result, lastUsr))
     }
     return result
-  }, [selectedUsers, selectedProjects, isFetching])
+  }, [selectedUsers, selectedProjects])
 
   const onSelectionChange = (e) => {
-    if (!onSelectUsers) return
+    if (!setSelectedUsers) return
     let result = []
     for (const user of e.value) result.push(user.name)
-    onSelectUsers(result)
-  }
-
-  // CTX
-
-  const onDelete = async () => {
-    confirmDialog({
-      message: `Are you sure you want to delete ${selectedUsers.length} user(s)?`,
-      header: 'Delete users',
-      icon: 'pi pi-exclamation-triangle',
-      accept: async () => {
-        for (const user of selectedUsers) {
-          try {
-            await axios.delete(`/api/users/${user}`)
-            toast.success(`Deleted user ${user}`)
-            refetch()
-          } catch {
-            toast.error(`Unable to delete user ${user}`)
-          }
-        }
-      },
-      reject: () => {},
-    })
+    setSelectedUsers(result)
   }
 
   // IDEA: Can these go into the details panel aswell?
@@ -155,67 +98,14 @@ const UserList = ({ selectedProjects, selectedUsers, onSelectUsers, setUserDetai
     },
   ]
 
-  let userListWithRoles = useMemo(
-    () => userList.map((user) => formatRoles(user, selectedProjects)),
-    [userList],
-  )
-
-  const searchableFields = ['name', 'attrib.fullName', 'attrib.email', 'rolesList', 'hasPassword']
-
-  const [search, setSearch, filteredData] = useSearchFilter(searchableFields, userListWithRoles)
-
   // Render
 
   return (
     <Section>
-      <ConfirmDialog />
-      <Toolbar>
-        <Button onClick={() => setShowNewUser(true)} label="Add a new user" icon="person_add" />
-        <Button
-          onClick={onDelete}
-          label="Delete selected users"
-          icon="person_remove"
-          disabled={!selectedUsers.length}
-        />
-        <InputText
-          style={{ width: '200px' }}
-          placeholder="Filter users..."
-          value={search}
-          onChange={(e) => setSearch(e.target.value)}
-        />
-      </Toolbar>
-
-      {showNewUser && (
-        <NewUserDialog
-          rolesList={rolesList}
-          onHide={() => {
-            setShowNewUser(false)
-          }}
-        />
-      )}
-
-      {showRenameUser && (
-        <RenameUserDialog
-          selectedUsers={selectedUsers}
-          onHide={() => {
-            setShowRenameUser(false)
-          }}
-        />
-      )}
-
-      {showSetPassword && (
-        <SetPasswordDialog
-          selectedUsers={selectedUsers}
-          onHide={() => {
-            setShowSetPassword(false)
-          }}
-        />
-      )}
-
       <TablePanel loading={isLoading || isLoadingRoles}>
         <ContextMenu model={contextMenuModel} ref={contextMenuRef} />
         <DataTable
-          value={filteredData}
+          value={tableList}
           scrollable="true"
           scrollHeight="flex"
           dataKey="name"
@@ -224,7 +114,7 @@ const UserList = ({ selectedProjects, selectedUsers, onSelectUsers, setUserDetai
           onContextMenu={(e) => contextMenuRef.current.show(e.originalEvent)}
           onContextMenuSelectionChange={(e) => {
             if (!selectedUsers.includes(e.value.name)) {
-              onSelectUsers([...selection, e.value.name])
+              setSelectedUsers([...selection, e.value.name])
             }
             setLastSelectedUser(e.data)
           }}
